Break up layout components into their own files
