+++ client/Layout/InitiativeListHost.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
-import { TrackerViewModel } from "./TrackerViewModel";
-import { useSubscription } from "./Combatant/linkComponentToObservables";
-import { InitiativeList } from "./InitiativeList/InitiativeList";
-import { CommandContext } from "./InitiativeList/CommandContext";
+import { TrackerViewModel } from "../TrackerViewModel";
+import { useSubscription } from "../Combatant/linkComponentToObservables";
+import { InitiativeList } from "../InitiativeList/InitiativeList";
+import { CommandContext } from "../InitiativeList/CommandContext";
 import { useCallback } from "react";
-import { TagState } from "../common/CombatantState";
+import { TagState } from "../../common/CombatantState";
 
 export function InitiativeListHost(props: { tracker: TrackerViewModel }) {
   const { tracker } = props;
